Marking a couple TODO and breaking api.go into api.go and api_test.go
